점프 수정 애니 수정
어택 충돌 설정 아직 없음
