+++ Assets/Scripts.zip.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 934d0bb6c65a6784eb0b77220c4d7b93
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
