--- Assets/Scripts/EnemyAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f84aff050e04724caeac6f967482dfd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
